sinput: add v2 feature descriptor
The current feature descriptor for sinput has some limitations that
will become apparent in the future. Since it is a new protocol, let us
introduce a V2 while it is still early. Specifically, V1 introduces
artificial limitations in how it encodes IMU data (e.g., calibration
information) and polling rate (you cannot use 800Hz; needs to be an even
number of ms). The feature flags are also limited (they are only two)
and the way touchpad support advertisement was added was not very
flexible.

This is not true with report 0x01, which is pretty standard and should
not need changing in the future.