sinput: add handheld type
Handhelds are supposed to be treated differently by consumers, especially
when it comes to gyro aiming. This is due to how the user holds the
device with the screen facing them. This causes a reversion of e.g.,
the roll axis, and requires a different default translation setting.

Therefore, add a handheld type so that consumers know how to setup gyro
aiming by default.