sinput: add feature report support for capabilities
Currently, sinput relies on writes and reads over the output bus to get
capabilities. This requires e.g., multiple retries and is timing based.
In addition, there is no check to ensure the controller is valid.

Both of these are error prone and prevent the protocol from proding
ambiguous controllers. Therefore, add a firm get report and set report
to force devices to identify themselves. First, the set report tells
the device this is the SINPUT protocol, and the device has the ability
to refuse that. Then, the in get report the device needs to respond
with 0x02, 0x02, 'S', 'I' for the report to be considered valid.